// src/generated/zod/inputTypeSchemas/AccountOrderByWithAggregationInputSchema.ts
import type { Prisma } from '@/generated/prisma'

import { z } from 'zod'
import { SortOrderSchema } from './SortOrderSchema'
import { SortOrderInputSchema } from './SortOrderInputSchema'
import { AccountCountOrderByAggregateInputSchema } from './AccountCountOrderByAggregateInputSchema'
import { AccountMaxOrderByAggregateInputSchema } from './AccountMaxOrderByAggregateInputSchema'
import { AccountMinOrderByAggregateInputSchema } from './AccountMinOrderByAggregateInputSchema'

export const AccountOrderByWithAggregationInputSchema: z.ZodType<Prisma.AccountOrderByWithAggregationInput> =
	z
		.object({
			id: z.lazy(() => SortOrderSchema).optional(),
			accountId: z.lazy(() => SortOrderSchema).optional(),
			providerId: z.lazy(() => SortOrderSchema).optional(),
			userId: z.lazy(() => SortOrderSchema).optional(),
			accessToken: z
				.union([
					z.lazy(() => SortOrderSchema),
					z.lazy(() => SortOrderInputSchema),
				])
				.optional(),
			refreshToken: z
				.union([
					z.lazy(() => SortOrderSchema),
					z.lazy(() => SortOrderInputSchema),
				])
				.optional(),
			idToken: z
				.union([
					z.lazy(() => SortOrderSchema),
					z.lazy(() => SortOrderInputSchema),
				])
				.optional(),
			accessTokenExpiresAt: z
				.union([
					z.lazy(() => SortOrderSchema),
					z.lazy(() => SortOrderInputSchema),
				])
				.optional(),
			refreshTokenExpiresAt: z
				.union([
					z.lazy(() => SortOrderSchema),
					z.lazy(() => SortOrderInputSchema),
				])
				.optional(),
			scope: z
				.union([
					z.lazy(() => SortOrderSchema),
					z.lazy(() => SortOrderInputSchema),
				])
				.optional(),
			password: z
				.union([
					z.lazy(() => SortOrderSchema),
					z.lazy(() => SortOrderInputSchema),
				])
				.optional(),
			createdAt: z.lazy(() => SortOrderSchema).optional(),
			updatedAt: z.lazy(() => SortOrderSchema).optional(),
			_count: z.lazy(() => AccountCountOrderByAggregateInputSchema).optional(),
			_max: z.lazy(() => AccountMaxOrderByAggregateInputSchema).optional(),
			_min: z.lazy(() => AccountMinOrderByAggregateInputSchema).optional(),
		})
		.strict()

export default AccountOrderByWithAggregationInputSchema
